+++ addon/prefs.js
@@ -1,3 +1,4 @@
 pref("extensions.annotationMarkdown.enabled", true);
 pref("extensions.annotationMarkdown.fontScalePercent", 100);
 pref("extensions.annotationMarkdown.pasteAsPlainText", true);
+pref("extensions.annotationMarkdown.mathEnabled", true);